requirements.txt
fastapi
python-multipart
uvicorn
jinja2
sqlalchemy
pandas
numpy
python-dotenv
langchain-community 
langchainhub 
langgraph
langchain[groq]
faiss-cpu
nltk
sentence_transformers
langchain-google-genai